Escape values to be displayed in token labels
If the value of a jquery.token source contained either { or } the
highlight regex would fail with an exception as documented here
 https://github.com/loopj/jquery-tokeninput/issues/785

Since the upstream project seems to be abandoned, making the fix in our
vendored version of the library. Also updating to the most recent verion
of the active admin 1.0.x release candidate.

It seems like the specs in this project have been broken for a while,
including the updated dummy db schema as a step towards getting those
straightened out.